Changed Behaviour to be Specification Compliant
Having read:
https://specifications.freedesktop.org/desktop-entry-spec

I discovered a "NoDisplay" tag in the .desktop file. It turns out, that I can just use this to remove files that shouldn't be there, and thus solve a need to check the mime-types to make sure I'm using the right menu item. 

That'll teach me to start without reading the specification.