Added a switch to time menu loading
After getting concerned that the changes made to program were slowing down the menu, I added a tag to show the load time of the menu, which generally comes in under 0.4s. 

This may mean that the menu load time may very well come down to the wait-time in openbox menu, which adds a pause before the menu loads.